Moved Gradle plugin releasing code to appropriate package
