+++ src/main/groovy/org/shipkit/internal/gradle/plugin/PluginDiscoveryPlugin.java
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle;
+package org.shipkit.internal.gradle.plugin;
 
 import com.gradle.publish.PluginBundleExtension;
 import com.gradle.publish.PluginConfig;
