+++ src/main/resources/META-INF/gradle-plugins/org.shipkit.gradle-plugin.properties
@@ -1,1 +1,1 @@
-implementation-class=org.shipkit.internal.gradle.GradlePluginReleasingPlugin
+implementation-class=org.shipkit.internal.gradle.plugin.GradlePluginReleasingPlugin
